requirements.txt
tensorflow-gpu
sklearn
transformers